Add "Accept all files" option for incoming files via BT [3/3]
* Adapted for DU-M.
* PS2: empty line fix
* PS3: variable name fix

BT server in AOSP will check MIME type of incoming file. Only those with
explicitly allowed types will be accepted. The MIME type whitelist is in
packages/apps/Bluetooth/src/com/android/bluetooth/opp/Constants.java
But this can be annoying when we want to transfer RAR, APK or conf files
etc which do not exist in the list.
This patch adds the option "Accept all files" in
Settings -> Bluetooth -> menu.
This option is disabled by default, which ensures security.

Change-Id: Idd2c438b4d0c0c6392e572e67b3ee9e8f9625282
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>