Added modded marginov color picker
-- Adapt for Preference being part of support library

@bigrushdog: ColorPickerPreference: Introduce set default button when attribute is set
Set a default value using the "defaultColorValue" settings namespace
attribute. The general idea is often we want set a value of say "-1"
and let framework handle it. One such scenario would be if you are
changing the color of a ImageView using ColorFilter. When framework
receives a "-1" value, framework sets the ImageView ColorFilter to null,
restoring the ImageView to original state.

@ezio84: Improve the code
* Allow tap also on the text to show the picker
* Don't show keyboard on start
* Don't show picker on click if preference is disabled
* Use oval preview
* Refresh color preview on click
* Change color picker reset icon

Change-Id: Ic721fdc7f63607d972d203e1a13097e8870bce2f