Allow unlocked USB data access (2/2)
by @KreAch3R

This bypasses the marshmallow "regression" of blocking and resetting USB
data access
after disconnecting. It also allows saving the default configuration and
it survives a reboot.
Originally, this "feature" was added by Google for security reasons
so it should be advised to use the below setting under caution.

Also:
* Adds an "Advanced" section under Storage and USB
* Moves Usb Configuration from Development settings
* Secured lockscreen disables the option

CypherOS Edits:
- Since storage manager is a preference screen in the volume settings,
let's remove the menu item. Replace the 3 dots with a connection icon
and once pressed by the user, this will launch USB Data Access.

Todo: Redo the UI to make it a bit more UF

Change-Id: I82369ec0af11796fd6b84aabe7ed1034aaec5ecf
Signed-off-by: xlxfoxxlx <joshfox87@gmail.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>