fingerpint: Allow devices to configure sensor location
The text for enrolling a fingerprint always assumes the sensor is
located on the back.  This patch adds a config integer that defines
where the sensor is located on the device.
  0 = back
  1 = front
  2 = left side
  3 = right side

Conflicts :
cm_strings.xml
configs.xml

This patch also adds a new drawable that shows the front of the device
when applicable.

Change-Id: Ia8f5f8ff4ba4d13d35aed3052f60ff665dd4f294
TICKET: CYNGNOS-3089