Settings: Forward port Gestures Settings
Brings back the Gestures menu accessible from the dashboard, which all
of the supported device gestures in one place.

- Move Tap to Wake to Gestures
- Move Lift to Wake to Gestures
- Move Double tap power for camera to Gestures and rebrand the option to
Double-snap with a new summary
- Utilize summary loader for to notify a user when Double-snap is
enabled or not
- Place the Gestures menu under sounds menu
- Add a new icon for Gestures menu
- Make Gestures menu searchable

To do: ensure Gestures is placed in the correct place in the device
category. Google's new dashboard placement is a bitch to get around.

Signed-off-by: Chris <sirhc95@cypheros.co>