Settings: Generate our own ReleaseID's for builds
This is purely for OTA updates to target this releaseid instead of just
the version alone. This is also handy in keeping track of newer versions
of the same build, like Release Canidates.

For example, we release a new build where the version number remains the
same, but the release id is updated to signify the newer build was
recieved. The last 4 digits are incremental while the first 4 are static
internal build indentifiers. The ReleaseID is also hardcoded for better
compatability with exporting it on build, to ensure it never displays
"Unspecified" unless we tell it to.

Todo: Cota needs to be reprogrammed to include these ReleaseID's in it's
scripts. Sooner or later, maybe we can generate this so that each device
can have it's on unique identifier and duplicate true OTA methods, that
generally use IMEI's as identifiers for staged rollouts