Settings: Battery styles [1/2]
This is our take on battery styles per UX basis. We've generated
a clean and simple interface for swicthing, what we call, battery views.
Users will be presented with a battery icon on the actionbar in battery settings.
Once tapped, a submenu will be generated with the proper battery views.

Change-Id: I5e9a547eab11fa17082392a8c6edcde0b8dd9665
Credits: @darkobas for his initial Settings implementation
Credits: @ParanoidAndroid for the idea of submenus for battery styles