Settings: Port LiveDislay
This commit ports over LiveDisplay from CMParts to
the native android Settings. Dependencies have been
refractored to work with the Settings app. We dropped
support for CMParts awhile ago, so this was needed.

- Target settings intervalSeekbar
- Add CMSystemSettingsSwitchPreference
- Add CustomDialogPreference
- Redirect xmls to utilize native directories
- Clean up strings and remove hardcoded title
- Add it to DisplaySettings and hide night display
- Remove night mode (Don't know why that was there)

Full credits go to @cyanogen for the initial implementation
Note: Proper CM dependencies must be acquired to have this working, such
as commits in frameworks/base and CMSDK must be integrated.

Change-Id: I21dd6c254407bf4ded1ce33a03cda96bc7ae9861