Settings: Add Team & Maintainer information
A new fragment that holds information on the CypherOS team
members and it's official maintainers. This list includes supported
devices by that individual maintainer. 

PS4: Completely rebased for UI/UX satisfaction. This information
deserves it's own fragment. Let's keep SoftwareInfo uncluttered.
Also generate a new G+ icon to differentiate profiles and the Project
community. Let's also resort the code to be organized for future
additions and make it easier to update.

Change-Id: Iaebd7687982ef893652e03ec3347539f409dfb1f