Gestures: New logic for launching external Motion Gestures
CMActions was rebased to support our common motion gestures.
Settings shall be updated to reflect this so it's launching
the correct class. Also new logic has been added to only show
the preference if the intent matches in it's parent activity,
since certain devices build Motion Gestures, it will only
initiate this for those devices

Change-Id: I05d7435c94c56f3e9a5af4dda2016374bdb68936