Settings: Move code for lockscreen notifications
DO NOT MERGE: This needs some clean up. The first patchset is a bit sloppy due
to my aggravation of having to do this. Lockscreen notifications have been moved
to Sound settings with the rest of the notification options. The dropdownpreference
doesn't work, mainly because it's missing code to initiate it. Code has been moved from
the ConfigureNotification class into the SoundSettings class. The AOSP notification
pulse code has been removed due to the new implementation.

Todo: Merge over LockscreenNotifications for Profile state. This will probably cause the whole
class to be removed anyway.

Change-Id: I48dc5aee04a36c4941bb2995bdbe0ed5a44c8175