Settings: Broker out Gestures into a custom class
This allows for Google's gestures to be used for custom
gesture features, such as DT2S or DT2W etc. The class
is moved into our common house and tweaked a bit to allow
custom preferences. Instead of allocating 2 separate classes
in our previous releases, merging it all into one, saves
space and it just makes sense.

Also use the notifications icon for the Sound & notifications
class on the dashboard.

Change-Id: I01ad4344633a26ec9b05b5477bde03cb109d57c4