Settings: Fix misconfigured actionbar menu items
ManageApplications: In the application settings the settings
gear menu item wasn't launchable. This is because we removed
it's menu case along with the configureapplications preference.
Let's add it back.

PowerUsageSummary: Due to the battery styles being implemented,
the menu ordering got messed up. The refresh icon launches battery
optimization instead. This is because it was labeled as "1" and so
was the optimization preference, making the system think it was the
same thing. Let's move all other preferences down a number so that
the refresh button can maintain it's order.

Change-Id: Id48fc811410da942d297b68bfc8441edb079e23f