Settings: Update battery styles menu drawable
The battery icon was a good approach, but it doesn't stand
out to indicate that something behind it can be configurable
for the UI.

Let's update the icon to a small paint brush, which
will signal that something is available to change. This should
give users a better idea on where the battery styles menu is
actually located.

Change-Id: Ib403c828a5978f673b74b8534c82b0b17f8672c5