Settings: Rewrite Carrier Label
This mainly rewrites the Edit Text preference dialog
to fit UI/UX standards. The text field is clipped to
fit perfectly inside the dialog.

An example is also given by default for users to give a
better idea of what can be placed as their network name,
which is generally anything they want. The disabling logic
has also be rewritten to work as intended. Previously, it
only initiates after the dialog is brought up and left empty.
This is poor UX.

Todo: Generate a new simple UI and format the preferences correctly,
we want this to blend natively in the settings menu.

Change-Id: Ic99a29282e9bc9fe7f6fa3a57828633a4358c495