DisplaySettings: Introduce Wakeup Protection
A rebranded fork of "Prevent accidental wakup".

How it works is, the device checks the proxmity sensor before waking the
screen to reduce accidental touches, unlocking and even butt dialing.
This feature is most useful while the device is in the users pocket.

Ported from: LineageOS/android_packages_apps_Settings@5aa2f34

Rewritten to be natively placed in display settings and not depend on
third party frameworks. Maintainers can configure the support bool as
well as it's enablement via overlay

Change-Id: I3659c70d595a62b39ea6a534fe54150d13c3cb21