ButtonBacklightBrightness: Update and refractor
CustomPreferenceDialog is an Android dialog to extendable custom
preferences. We had our own and left the one in the settings main folder
untouched. But after some review the native CustomPreferenceDialog get's
properly initiated with onDisplayPreferenceDialog, which is a required
override fragment based dialogs.

So let's remove our modified dialog preference and move some of the code
out into the native one just to extend support for more explicit
dialogs.

This should fix the ButtonBacklight crashes that's been happening for
weeks

Change-Id: Ib094542152d8f8fa4c95f0f376a7ba27525315d1