ButtonBacklightBrightness: Import UnrestrictedDialogPreference
CustomDialogPreference is an Android dialog to extendable custom
preferences. We had our own and left the one in the settings main folder
untouched. But, our's caused crashes on RT due to onDisplayPreferenceDialog
not being overidden. After review on how the CustomDialogPreference API
worked, the best solution was to create a new dialog preference, place
it in the main settings with the rest and initiate it's preference key
from SettingsPreferenceFragment which allows onDisplayPreferenceDialog
to be overriden properly.

Our UnrestrictedDialogPreference is referenced from the original Android
CustomDialogPreference with added fucntions pulled from Lineage's
CustomDialogPreference, which they use in their CMParts application. Since
the buttons implementation is ported from CMParts, some dialog functions
must follow for ButtonBacklightBrightness.

This should fix the ButtonBacklight crashes that's been happening for
weeks

Change-Id: Ib094542152d8f8fa4c95f0f376a7ba27525315d1