Settings: Remove navbar animation
The 7.1.2 merge made several changes to the way the navbar works, which
broke orientation for the navbar animation. Our fixes caused systemui
crashes when the home button is pressed, meaning this needs some
refractoring. Until then, let's remove the option to toggle.

This will be disabled by default in base

Change-Id: Id857b3e5316a3a9a708a3e45a5fbef8bebfd19b9