Introduce Support Manager Interface
This adds a support manager on the settings dashboard in the form
of a slidingtab. Users can swipe to the secondary dashboard which
will provide software based support such as feature locations,
tips and tricks, information about our software and a dedicated
bug reporting option.

Our support manager is based on Google's support feature, introduced
in nougat for Pixel Devices. The infrastructure is similar, but we
don't offer chat and phone support, so let's make use of this with for
something we offer.

DO NOT MERGE: Google's support interface is based on a feature overlay,
that
must be enabled properly. We intentionally stripped parts from it and
combined
it with our own implementation. First objective is to confirm this shows
as
a proper slidingtab in the dashboard, then we continue with layout
construction.

PS3: Rebase Support Manager on the latest AOSP changes. PS1 was heavily outdated
from last year

Change-Id: Ic09214c81b1f91d938e4e52eedc04841a39e0612