SupportManager: Add new Weblinks item
Weblinks are the new shortcut based interface, rebased on
software info in the "About Phone" settings. Users can
get direct access to weblinks such as our Google or Github
source page right from Support Manager.

So far, it only launches a fragment called as an activity, but
the real concept is a custom dialog interface with a weblink
chooser.

Change-Id: I295bd0420eaffdc1cfdbb25fb5fb47a4b68d5508