SupportManager: Add bug report button item
This adds a button item under the welcome greeting. When the
bug report button is clicked a dialog will show with bug reporting
options. This is the native AOSP bug reporting feature and will
be used until we create our own.

Note: The initial commit is only a placeholder to confirm everything
is working. Once verified, this placeholder will be in the next update
and a patch will be applied once available to open up bug reporting
officially. This placeholder only creates the feel of Support Manager.

When the user presses the button, they'll be greeted with a toast
acknowledging
that bug reporting is unavailable.

Change-Id: I2ebe785f45bdea9f861ef8e6a03d07a9b4c8ba95