SupportManager: Add email button item
This adds a button item under the welcome greeting. When
the email button is clicked, email dialog or client will
be launched to allow the user to email the team for a subset
of reasons

Note: The initial commit is only a placeholder to confirm everything
is working. Once verified, this placeholder will be in the next update
and a patch will be applied once available to open up the email client
officially. This placeholder only creates the feel of Support Manager.

Also added case if Support button actions are unavailable somehow, or
disabled by a maintainer

Change-Id: I81b98c17e7ec5492498d99a10716d5a032dd7238