SupportManager: Add email client functionality
Allows the user to email the team with one tap
from the Support Manager. This allow's users to
give feedback about the rom as well as suggest
new ideas.

The current implementation uses a share chooser
for available email clients. If there's no supported
email client available, the dialog will inform the
user.

In the future, we hope to create a full functional
remote dialog to parse emails and send without the
need to open the email client.

Change-Id: Iecb1ce796c7de99fa5a465fcba6e41e2f804f8ad