Settings: add home button animation setting
Allows user to change the state of the home button
animation with new logic. The preference is set based
on the supported config. If the config is set to true
the preference will appear in button settings enabled
by default. If the config is set to false the preference
will not appear in the settings at all.

This follows config support in frameworks/base to not
allow a switch with no supported system state.

Change-Id: Ib513f5a8a77766a0dc21bdf9476bf76551526264