Settings: add TwoChoicePreference
The logic is simple, it's a basic custom preference that allows to
choices for the user.

1. The ability to launch the preference into a new screen (PreferenceScreen)
2. The ability to change the state of the preference without entering
the preference. It's generally a SwitchPreference and PreferenceScreen
combined.

This is similar to the old WiFiEnabler (etc.) switches on the dashboard but
adapted to work with generic PreferenceScreens. This design follows the
basic switch enablers but with a divider that separates the two to give
the user an indication that there are two choices. Google uses this logic,
heavily in O, and i feel it's very useful instead of converting a switch
preference into a preference screen.

Change-Id: Ifb35e1341f1d3df3b3c8f9bd5a057f1f7546df0a