Buttons: allow user to enable/disable animation [1/2]
Adds a preference to enable/disable the navigation bar
animation. Off by default.

Written for multiple usage cases. Preference only shows if the
opaLayout config is set to "true" on build. If the config is set
to true, the preference then checks if the navigation bar is enabled
before allowing enable/disable functionality, otherwise it will disable.
This was set for hw key devices that can disable the navigation bar
entirely.

Change-Id: I38ad8d81900078b5d888c170796d78004d7131ce