Buttons: fix availability checks for navbar animation
Null checks can't be initiated on preferences that aren't even
available. isAvailable checks for availability from the start. This
causes the preference to not show, regardless of the config's status.
Simply make it's return value dependent on the config's status.

Note to self: Stop writing code at 4 in the morning, you screw up!

Change-Id: If869a9f62bf6d947b01e906ac8c0715b5d0e909f