ColorOM: Logical Upgrade to 2.0 [1/2]
Other half of https://gerrit.cypheros.co/#/c/CypherOS/frameworks_base/+/3087/

Make accents be more formal with base themes. Lot's of workaround have been
done here for the radio buttons to work properly. Currently, the ammount of
code is too much for my taste, we'll shorten this later.

Temporarily disable the Custom Accent chooser until we remove the default
option and rewrite it to the one we have available.

Change-Id: Ieb44c7d8e7e305a76557e2ea936eaa95cd65ee3a