ColorOM: Fix issue where accents can't apply
Sometimes accents won't apply. We use "setEnabled" instead
of "setEnabledExclusive" so that's why, assuming sometimes overlaying
an existing overlay causes issues. To work around this, we should
quickly disable the existing accent before applying the new one.

This may cause a UI flicker, but it's the best solution for now until
a better one is found.

- Added a new central class for handling task based methods

Change-Id: Ic4be8ebedc9db61d7b0a0af390d6ff8d06243922