deviceinfo: Add support for overriding vendor properties
Some properties will forcefully be read from vendor instead of system
which can result in outdated or misrepresented OEM information. As a
small workaround, override the current props that are displayed in about
phone if the config is set to true.

This way we can set our own props and display proper information such as
the security patch level. Although this is minor, asthetics matter.
Current supported props are

"ro.vendor.override.model"
"ro.vendor.override.patch"

which override the devices model name and security patch level. This is
a last resort incase we can't override these traditionally

Change-Id: Ia00a23e45364e1546d5f5ebe55867b1c27340450