Improve Fingerprint Enroll UI
Improves the fingerprint enroll process with proper illustration
during each step that caters to a device with specific fingerprint
placement. So far, we support bottom sensors as well and illustrations
reflect this after adding the bottom sensor overlay.

Also, on devices with bottom sensors, we've added automatic capability
to disabling the navigation bar during the enrollment process to help
further the enrollment process.

Change-Id: I2489006d5550395890e6b38a55b00de950ffe695