deviceinfo: Implement Multi-User v2
Implements/Moves Multi-User access to the action bar in the
about phone menu. Previously, user information was displayed via
header with the actual settings in the root dashboard. This
combines it into one tap. 

The action bar displays the current user's profile icon and
user name (when in landscape). Tapping the icon will bring
in the traditional Multi-User settings

Change-Id: I5fc1c85f36682104b672a1e8fbe7c16063ed7d0a