deviceinfo: Add animatable aoscp version header
Implements the aoscp basic version information in it's own header on the
about phone root.

This is shown based on the multi-user version. In v1, we show the basic
user profile with aoscp version information bundled in the Android
dialog controller.

- Add 2 new local API's for LayoutPreference
- isCustom(boolean): Decides whether we're inflating a
custom layout in code, without defining layouts in xml.
This allows us to dynamically inflate layouts.
- setCustomLayout(int): Specifies the custom layout
being used. isCustom must be set the "true" to specify
a custom layout

Change-Id: I5f2d222b68d6da2425bc518317767db6a7d36abd