Add preference for Display Mode
Our custom fork of aosp's ColorModes feature which uses
aoscp hardware. If the device supports display modes, it
must overlay the DisplayEngineController with it's hw methods
and overlay the current placeholder strings with the mode names
conducted by their vendor.

Starting with this commit, we support up to 5 display modes and
the ammount of modes show based on the ammount declared in
DisplayEngineController with getAvailableModes()

todo: Add a function to inherit dynamic mode names from
      DisplayEngineController

Change-Id: I8453e4ef7586e795f7933eaa92ecad1983b46d8c